remove deprecated console differ, to use symplify
