--- packages/console-differ/src/Diff/Output/CompleteUnifiedDiffOutputBuilderFactory.php
@@ -1,36 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\ConsoleDiffer\Diff\Output;
-
-use SebastianBergmann\Diff\Output\UnifiedDiffOutputBuilder;
-use Symplify\PackageBuilder\Reflection\PrivatesAccessor;
-
-/**
- * Creates @see UnifiedDiffOutputBuilder with "$contextLines = 1000;"
- */
-final class CompleteUnifiedDiffOutputBuilderFactory
-{
-    /**
-     * @var PrivatesAccessor
-     */
-    private $privatesAccessor;
-
-    public function __construct()
-    {
-        $this->privatesAccessor = new PrivatesAccessor();
-    }
-
-    /**
-     * @api
-     */
-    public function create(): UnifiedDiffOutputBuilder
-    {
-        $unifiedDiffOutputBuilder = new UnifiedDiffOutputBuilder('');
-
-        $this->privatesAccessor->setPrivateProperty($unifiedDiffOutputBuilder, 'contextLines', 1000);
-
-        return $unifiedDiffOutputBuilder;
-    }
-}
